refactor/#4: 회원가입 폴더 이름 변경(auth->user)
